Global parameter and local parameter calculation external expansion.
