+++ dolphinscheduler-master/src/main/java/org/apache/dolphinscheduler/server/master/runner/task/BaseTaskProcessor.java
@@ -73,6 +73,7 @@ import org.apache.dolphinscheduler.plugin.task.k8s.K8sTaskParameters;
 import org.apache.dolphinscheduler.server.builder.TaskExecutionContextBuilder;
 import org.apache.dolphinscheduler.server.master.config.MasterConfig;
 import org.apache.dolphinscheduler.service.bean.SpringApplicationContext;
+import org.apache.dolphinscheduler.service.expand.CuringParamsService;
 import org.apache.dolphinscheduler.service.process.ProcessService;
 import org.apache.dolphinscheduler.service.task.TaskPluginManager;
 import org.apache.dolphinscheduler.spi.enums.DbType;
@@ -123,6 +124,8 @@ public abstract class BaseTaskProcessor implements ITaskProcessor {
 
     protected TaskPluginManager taskPluginManager;
 
+    protected CuringParamsService curingParamsService;
+
     protected String threadLoggerInfoName;
 
     @Override
@@ -130,6 +133,7 @@ public abstract class BaseTaskProcessor implements ITaskProcessor {
         processService = SpringApplicationContext.getBean(ProcessService.class);
         masterConfig = SpringApplicationContext.getBean(MasterConfig.class);
         taskPluginManager = SpringApplicationContext.getBean(TaskPluginManager.class);
+        curingParamsService = SpringApplicationContext.getBean(CuringParamsService.class);
         this.taskInstance = taskInstance;
         this.processInstance = processInstance;
         this.maxRetryTimes = masterConfig.getTaskCommitRetryTimes();
@@ -301,6 +305,7 @@ public abstract class BaseTaskProcessor implements ITaskProcessor {
             setK8sTaskRelation(k8sTaskExecutionContext, taskInstance);
         }
 
+        AbstractParameters baseParam = taskPluginManager.getParameters(ParametersNode.builder().taskType(taskInstance.getTaskType()).taskParams(taskInstance.getTaskParams()).build());
         return TaskExecutionContextBuilder.get()
                 .buildTaskInstanceRelatedInfo(taskInstance)
                 .buildTaskDefinitionRelatedInfo(taskInstance.getTaskDefine())
@@ -309,6 +314,7 @@ public abstract class BaseTaskProcessor implements ITaskProcessor {
                 .buildResourceParametersInfo(resources)
                 .buildDataQualityTaskExecutionContext(dataQualityTaskExecutionContext)
                 .buildK8sTaskRelatedInfo(k8sTaskExecutionContext)
+                .buildParamInfo(curingParamsService, baseParam, processInstance)
                 .create();
     }
