+++ dolphinscheduler-service/src/main/java/org/apache/dolphinscheduler/service/expand/CuringParamsService.java
@@ -15,16 +15,19 @@
  * limitations under the License.
  */
 
-package org.apache.dolphinscheduler.common.expand;
+package org.apache.dolphinscheduler.service.expand;
 
 import org.apache.dolphinscheduler.common.enums.CommandType;
+import org.apache.dolphinscheduler.dao.entity.ProcessInstance;
+import org.apache.dolphinscheduler.plugin.task.api.TaskExecutionContext;
 import org.apache.dolphinscheduler.plugin.task.api.model.Property;
+import org.apache.dolphinscheduler.plugin.task.api.parameters.AbstractParameters;
 
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
 
-public interface CuringGlobalParamsService {
+public interface CuringParamsService {
 
     /**
      * time function need expand
@@ -61,4 +64,13 @@ public interface CuringGlobalParamsService {
      * @return
      */
     String curingGlobalParams(Integer processInstanceId, Map<String, String> globalParamMap, List<Property> globalParamList, CommandType commandType, Date scheduleTime, String timezone);
+
+    /**
+     * param parsing preparation
+     * @param baseParam
+     * @param taskExecutionContext
+     * @param processInstance
+     * @return
+     */
+    Map<String, Property> paramParsingPreparation(TaskExecutionContext taskExecutionContext, AbstractParameters baseParam, ProcessInstance processInstance);
 }
