+++ dolphinscheduler-service/src/test/java/org/apache/dolphinscheduler/service/expand/CuringGlobalParamsServiceTest.java
@@ -16,7 +16,7 @@
  */
 
 
-package org.apache.dolphinscheduler.common.expand;
+package org.apache.dolphinscheduler.service.expand;
 
 import org.apache.dolphinscheduler.common.enums.CommandType;
 import org.apache.dolphinscheduler.common.utils.DateUtils;
@@ -45,7 +45,7 @@ public class CuringGlobalParamsServiceTest {
     private static final String placeHolderName = "$[yyyy-MM-dd-1]";
 
     @Mock
-    private CuringGlobalParamsService curingGlobalParamsService;
+    private CuringParamsService curingGlobalParamsService;
 
     @InjectMocks
     private DolphinSchedulerCuringGlobalParams dolphinSchedulerCuringGlobalParams;
