+++ dolphinscheduler-task-plugin/dolphinscheduler-task-api/src/main/java/org/apache/dolphinscheduler/plugin/task/api/TaskExecutionContext.java
@@ -183,6 +183,11 @@ public class TaskExecutionContext implements Serializable {
     private Map<String, String> definedParams;
 
     /**
+     * prepare params map
+     */
+    private Map<String, Property> prepareParamsMap;
+
+    /**
      * task AppId
      */
     private String taskAppId;
