+++ dolphinscheduler-task-plugin/dolphinscheduler-task-api/src/main/java/org/apache/dolphinscheduler/plugin/task/api/parser/ParameterUtils.java
@@ -22,19 +22,15 @@ import static org.apache.dolphinscheduler.plugin.task.api.TaskConstants.PARAMETE
 import static org.apache.dolphinscheduler.plugin.task.api.TaskConstants.PARAMETER_SHECDULE_TIME;
 
 import org.apache.dolphinscheduler.plugin.task.api.model.Property;
-import org.apache.dolphinscheduler.spi.enums.CommandType;
 import org.apache.dolphinscheduler.plugin.task.api.enums.DataType;
 import org.apache.dolphinscheduler.spi.utils.DateUtils;
-import org.apache.dolphinscheduler.spi.utils.JSONUtils;
 import org.apache.dolphinscheduler.spi.utils.StringUtils;
 
 import java.sql.PreparedStatement;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.Iterator;
-import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -155,54 +151,59 @@ public class ParameterUtils {
     /**
      * curing user define parameters
      *
+     * Warning:
+     *      Change time: 2022-06-30
+     *      The purpose is for external expansion of local parameters.
+     *      now the method is replaced by the curingGlobalParams() method of DolphinSchedulerCuringGlobalParams.
+     *
      * @param globalParamMap  global param map
      * @param globalParamList global param list
      * @param commandType     command type
      * @param scheduleTime    schedule time
      * @return curing user define parameters
      */
-    public static String curingGlobalParams(Map<String, String> globalParamMap, List<Property> globalParamList,
-                                            CommandType commandType, Date scheduleTime) {
-
-        if (globalParamList == null || globalParamList.isEmpty()) {
-            return null;
-        }
-
-        Map<String, String> globalMap = new HashMap<>();
-        if (globalParamMap != null) {
-            globalMap.putAll(globalParamMap);
-        }
-        Map<String, String> allParamMap = new HashMap<>();
-        //If it is a complement, a complement time needs to be passed in, according to the task type
-        Map<String, String> timeParams = BusinessTimeUtils
-            .getBusinessTime(commandType, scheduleTime);
-
-        if (timeParams != null) {
-            allParamMap.putAll(timeParams);
-        }
-
-        allParamMap.putAll(globalMap);
-
-        Set<Map.Entry<String, String>> entries = allParamMap.entrySet();
-
-        Map<String, String> resolveMap = new HashMap<>();
-        for (Map.Entry<String, String> entry : entries) {
-            String val = entry.getValue();
-            if (val.startsWith("$")) {
-                String str = ParameterUtils.convertParameterPlaceholders(val, allParamMap);
-                resolveMap.put(entry.getKey(), str);
-            }
-        }
-        globalMap.putAll(resolveMap);
-
-        for (Property property : globalParamList) {
-            String val = globalMap.get(property.getProp());
-            if (val != null) {
-                property.setValue(val);
-            }
-        }
-        return JSONUtils.toJsonString(globalParamList);
-    }
+//    public static String curingGlobalParams(Map<String, String> globalParamMap, List<Property> globalParamList,
+//                                            CommandType commandType, Date scheduleTime) {
+//
+//        if (globalParamList == null || globalParamList.isEmpty()) {
+//            return null;
+//        }
+//
+//        Map<String, String> globalMap = new HashMap<>();
+//        if (globalParamMap != null) {
+//            globalMap.putAll(globalParamMap);
+//        }
+//        Map<String, String> allParamMap = new HashMap<>();
+//        //If it is a complement, a complement time needs to be passed in, according to the task type
+//        Map<String, String> timeParams = BusinessTimeUtils
+//            .getBusinessTime(commandType, scheduleTime);
+//
+//        if (timeParams != null) {
+//            allParamMap.putAll(timeParams);
+//        }
+//
+//        allParamMap.putAll(globalMap);
+//
+//        Set<Map.Entry<String, String>> entries = allParamMap.entrySet();
+//
+//        Map<String, String> resolveMap = new HashMap<>();
+//        for (Map.Entry<String, String> entry : entries) {
+//            String val = entry.getValue();
+//            if (val.startsWith("$")) {
+//                String str = ParameterUtils.convertParameterPlaceholders(val, allParamMap);
+//                resolveMap.put(entry.getKey(), str);
+//            }
+//        }
+//        globalMap.putAll(resolveMap);
+//
+//        for (Property property : globalParamList) {
+//            String val = globalMap.get(property.getProp());
+//            if (val != null) {
+//                property.setValue(val);
+//            }
+//        }
+//        return JSONUtils.toJsonString(globalParamList);
+//    }
 
     /**
      * $[yyyyMMdd] replace schedule time
